requirements.txt
joblib==0.14.1
numpy==1.18.3
paho-mqtt==1.5.0
pandas
pathlib
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.16
